+++ CHANGELOG.rst
@@ -1,6 +1,9 @@
 3.2.7 (unreleased)
 ------------------
 
+- Be able to customize how you create es connection object
+  [vangheem]
+
 - Not loading utility by default and avoid crash when utility is not configured
   [bloodbare]